Aplicação regras de acesso aos endpoints

+++ services/autor.service.js
@@ -1,4 +1,5 @@
 import AutorRepository from "../repositories/autor.repository.js"
+import LivroRepository from "../repositories/livro.repository.js";
 
 async function createAutor(autor){
     return await AutorRepository.insertAutor(autor);
@@ -12,11 +13,12 @@ async function getAutor(id){
     return await AutorRepository.getAutor(id);
 }
 
-async function deleteAutor(id){
-    // if(await AlunoRepository.getAlunos(id).length){
-    //     throw new Error('Não foi possível excluir o Autor informado. Existem aluno(s) associado(s) a ele.')
-    // }
-    //todo: antes de excluir um autor, verificar se existem livros cadastrados para ele. Caso exista, bloquear a exclusão
+async function deleteAutor(id){    
+    const livros = await LivroRepository.getLivros(id);
+    
+    if(livros.length){
+        throw new Error('Não foi possível excluir o Autor informado. Existem livros(s) associado(s) a ele.')
+    }
     await AutorRepository.deleteAutor(id);
 }
